# frontend/lib/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "allwin API",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/leagues": {
   "get": {
    "tags": [
     "public"
    ],
    "summary": "List Leagues",
    "operationId": "list_leagues_api_v1_leagues_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/LeagueInfo"
         },
         "type": "array",
         "title": "Response List Leagues Api V1 Leagues Get"
        }
       }
      }
     },
     "400": {
      "description": "Bad Request",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "401": {
      "description": "Unauthorized",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "403": {
      "description": "Forbidden",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "404": {
      "description": "Not Found",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "422": {
      "description": "Unprocessable Content",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/leagues/{league_id}/standings": {
   "get": {
    "tags": [
     "public"
    ],
    "summary": "League Standings",
    "operationId": "league_standings_api_v1_leagues__league_id__standings_get",
    "parameters": [
     {
      "name": "league_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "League Id"
      }
     },
     {
      "name": "season",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Season"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/StandingsResponse"
        }
       }
      }
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Bad Request"
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unauthorized"
     },
     "403": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Forbidden"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Not Found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    }
   }
  },
  "/api/v1/leagues/{league_id}/fixtures": {
   "get": {
    "tags": [
     "public"
    ],
    "summary": "League Fixtures",
    "operationId": "league_fixtures_api_v1_leagues__league_id__fixtures_get",
    "parameters": [
     {
      "name": "league_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "League Id"
      }
     },
     {
      "name": "season",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Season"
      }
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "pattern": "^(upcoming|finished)$"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 400,
       "minimum": 1,
       "default": 100,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MatchListResponse"
        }
       }
      }
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Bad Request"
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unauthorized"
     },
     "403": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Forbidden"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Not Found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    }
   }
  },
  "/api/v1/matches": {
   "get": {
    "tags": [
     "public"
    ],
    "summary": "List Matches",
    "description": "比赛列表:只返回请求者有权限的联赛(免费=英超)。",
    "operationId": "list_matches_api_v1_matches_get",
    "parameters": [
     {
      "name": "date",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
        },
        {
         "type": "null"
        }
       ],
       "title": "Date"
      }
     },
     {
      "name": "league_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "League Id"
      }
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "pattern": "^(upcoming|finished)$"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MatchListResponse"
        }
       }
      }
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Bad Request"
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unauthorized"
     },
     "403": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Forbidden"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Not Found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    }
   }
  },
  "/api/v1/matches/{match_id}": {
   "get": {
    "tags": [
     "public"
    ],
    "summary": "Match Detail",
    "operationId": "match_detail_api_v1_matches__match_id__get",
    "parameters": [
     {
      "name": "match_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Match Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MatchDetailResponse"
        }
       }
      }
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Bad Request"
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unauthorized"
     },
     "403": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Forbidden"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Not Found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    }
   }
  },
  "/api/v1/matches/{match_id}/prediction": {
   "get": {
    "tags": [
     "public"
    ],
    "summary": "Match Prediction",
    "description": "免费/匿名:只有最高一项概率(受限字段物理不存在);Pro/Premium:完整 WDL。\n\n响应随 entitlement 变化 → 永远 private, no-store,禁止共享缓存。",
    "operationId": "match_prediction_api_v1_matches__match_id__prediction_get",
    "parameters": [
     {
      "name": "match_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Match Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PredictionResponse"
        }
       }
      }
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Bad Request"
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unauthorized"
     },
     "403": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Forbidden"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Not Found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "ApiErrorDTO": {
    "properties": {
     "code": {
      "type": "string",
      "title": "Code"
     },
     "message": {
      "type": "string",
      "title": "Message"
     },
     "details": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Details"
     }
    },
    "type": "object",
    "required": [
     "code",
     "message",
     "details"
    ],
    "title": "ApiErrorDTO"
   },
...
   "LeagueInfo": {
    "properties": {
     "league_id": {
      "type": "integer",
      "title": "League Id"
     },
     "code": {
      "type": "string",
      "title": "Code"
     },
     "name_zh": {
      "type": "string",
      "title": "Name Zh"
     },
     "name_en": {
      "type": "string",
      "title": "Name En"
     },
     "entitlement": {
      "type": "string",
      "title": "Entitlement"
     },
     "accessible": {
      "type": "boolean",
      "title": "Accessible"
     }
    },
    "type": "object",
    "required": [
     "league_id",
     "code",
     "name_zh",
     "name_en",
     "entitlement",
     "accessible"
    ],
    "title": "LeagueInfo"
   },
...
   "MatchListResponse": {
    "properties": {
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "limit": {
      "type": "integer",
      "title": "Limit"
     },
     "offset": {
      "type": "integer",
      "title": "Offset"
     },
     "matches": {
      "items": {
       "$ref": "#/components/schemas/MatchSummary"
      },
      "type": "array",
      "title": "Matches"
     }
    },
    "type": "object",
    "required": [
     "total",
     "limit",
     "offset",
     "matches"
    ],
    "title": "MatchListResponse"
   },
...
   "StandingsResponse": {
    "properties": {
     "league_id": {
      "type": "integer",
      "title": "League Id"
     },
     "season": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Season"
     },
     "available_seasons": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Available Seasons"
     },
     "rows": {
      "items": {
       "$ref": "#/components/schemas/StandingRow"
      },
      "type": "array",
      "title": "Rows"
     },
     "empty_reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Empty Reason"
     }
    },
    "type": "object",
    "required": [
     "league_id",
     "available_seasons",
     "rows"
    ],
    "title": "StandingsResponse",
    "description": "empty_reason 只在无数据时出现(端点用 response_model_exclude_unset 保持现状)。"
   },
...
  }
 }
}